+++ graphql/schema/user.js
@@ -28,7 +28,7 @@ const userTypeDefs = gql`
 const userResolvers = {
 	me: async (obj, args, context, info) => {
 		// Permittable by users
-		if (!context.user) {
+		if (!context.user.isAuth) {
 			throw new Error('User not logged in!');
 		}
 
@@ -65,7 +65,7 @@ const userResolvers = {
 
 	grantAdmin: async (obj, { email }, context, info) => {
 		// Permittable by admins
-		if (!context.user) {
+		if (!context.user.isAuth) {
 			throw new Error('User not logged in!');
 		}